requirements.txt
Flask==3.0.0
pydicom==2.4.4
requests==2.31.0
python-dotenv==1.0.1
